refactor: Move the serving primitives out of engine.py
`engine.py` is sixteen hundred lines and does several jobs: it builds images,
runs Tesseracts, parses volumes, and -- incidentally -- knows how to pick a port,
wait for an HTTP server to answer, and tell a lost race for a port apart from a
Tesseract that is genuinely broken.

That last group is one concern and not really about Docker at all, so it moves to
`serving.py`. `engine` imports it back, so nothing that referred to these by name
has to change; the tests that stub `requests.get` now stub it where the requests
are actually made.
